adjust image code function
