+++ web/src/routes/instance.go
@@ -670,21 +670,6 @@ func (a *InstanceAdmin) buildMetadata(ctx context.Context, primaryIface *Interfa
 	return interfaces, string(jsonData), nil
 }
 
-func GetImageOSCode(ctx context.Context, instance *model.Instance) string {
-	osCode := "linux"
-	if instance.Image == nil {
-		_, db := GetContextDB(ctx)
-		instance.Image = &model.Image{Model: model.Model{ID: instance.ImageID}}
-		err := db.Take(instance.Image).Error
-		if err != nil {
-			logger.Error("Invalid image ", instance.ImageID)
-			return osCode
-		}
-	}
-	osCode = instance.Image.OSCode
-	return osCode
-}
-
 func (a *InstanceAdmin) GetMetadata(ctx context.Context, instance *model.Instance, rootPasswd string) (metadata string, err error) {
 	vlans := []*VlanInfo{}
 	instLinks := []*NetworkLink{}
